+++ fuzz/mk_md4_fuzz.h
@@ -2,10 +2,7 @@
 #define mk_include_guard_md4_fuzz
 
 
-#include <stddef.h> /* size_t */
-
-
-void mk_md4_fuzz(unsigned char const* data, size_t size);
+void mk_md4_fuzz(unsigned char const* data, int size);
 
 
 #endif
